add comment and simplify code

+++ dali/kernels/test/simd_test.cc
@@ -49,8 +49,9 @@ template <typename In, int nvec = sizeof(float)/sizeof(In)>
 void TestConvertLoad(In lo, In hi) {
   int lanes = 16 / sizeof(In);
   In in[lanes];  // NOLINT
-  for (int i = 0; i < lanes; i++) {
-    in[i] = int64_t(lo) + (int64_t(hi) - int64_t(lo)) * i / (lanes - 1);
+  // promote to int64_t computation to avoid integer overflow
+  for (int64_t i = 0; i < lanes; i++) {
+    in[i] = lo + (hi - lo) * i / (lanes - 1);
   }
   float4x<nvec> v = load_f(in);
   for (int i = 0; i < nvec; i++) {
